arch: D-series — 5 more route groups extracted, codec_dashboard 3618→2921
Continuation of route extraction (PR #161 deferrals).
Pytest: 2148 → 2172 (+24 regression tests, 0 regressions).

ROUTES EXTRACTED

D1 / SR-42: /api/qchat/*       → routes/qchat.py
  5 endpoints (sessions/{sid}/save/{sid}/search) + QCHAT_DB + qchat_db()
  helper. SQLite WAL + busy_timeout + auto-migration of user_id column.

D2 / SR-43: /api/vibe/*        → routes/vibe.py
  4 endpoints (sessions/session/save/delete) + VIBE_DB + vibe_db() helper.
  Same lifecycle pattern as qchat.

D3 / SR-44: /api/schedules/*   → routes/schedules.py
  6 endpoints including the /run endpoint with its background-thread
  _execute_task() (LLM report → Google Doc → notification flow).
  PUT endpoint uses codec_jsonstore.atomic_write_json (re-audit medium
  fix preserved).

D4 / SR-45: /api/prompts/*     → routes/prompts.py
  3 endpoints + the 3 helper functions (_load_prompt_overrides,
  _save_prompt_overrides, _get_all_prompts). codec_dashboard's
  chat_completion handler now lazy-imports _load_prompt_overrides
  from routes.prompts to avoid a module-load-time cycle.

D5 / SR-46: media endpoints    → routes/media.py
  6 endpoints (webcam x3, screenshot, clipboard x2). Vision-proxy POST
  in webcam_capture added to test_a12_invariant._ALLOWLIST (legitimate
  vision site pending A-11 cleanup).

NOT INCLUDED IN D5
- /api/upload and /api/upload_image (250 LOC with B1 fence markers +
  size cap + path safety) — kept in codec_dashboard.py to keep the
  security-hardening review surface tight. Future PR.

LOC SHIFT

  codec_dashboard.py: 3,618 → 2,921 LOC (-697, -19.3%)
                     (started at 3,912 before PR #160)

NEW TESTS

  tests/test_route_extractions_d.py — 24 tests across 5 phases.
  Each phase parametrizes endpoint-registered checks + source-side
  invariants (db helpers in module, atomic-write in schedules, lazy
  prompt import in dashboard). Global LOC floor tightened to < 3,000.

UPDATED SOURCE-SCAN TESTS

- tests/test_a12_invariant.py: routes/media.py added to allowlist
  (vision POST in webcam_capture, A-11 pending)
- tests/test_dashboard_llm.py: /chat/completions count assertion
  bumped 5 → 4 (vision POST moved out with media)

Full pytest: 2172 passed, 83 skipped (env-dependent), 0 failed (1m24s).
