+++ tests/test_a12_invariant.py
@@ -28,6 +28,7 @@ _ALLOWLIST = {
     "skills/screenshot_text.py",  # OCR vision POST (A-11 pending)
     "codec_core.py",             # generated session-script string, not a live POST
     "pilot/pilot_agent.py",      # vendored CODEC Pilot module — uses its own LLM client
+    "routes/media.py",           # webcam vision POST (A-11 pending) — D5 extraction
 }
 
 _INLINE_POST_RE = re.compile(r"\.post\s*\([^)]*chat/completions")
