+++ tests/test_dashboard_llm.py
@@ -63,10 +63,12 @@ def test_dashboard_uses_codec_llm_call():
 
 def test_dashboard_only_vision_posts_remain():
     # After PR-3E-chat-stream migrated the chat stream + non-stream fallback,
-    # the ONLY inline /chat/completions occurrences left are the 5 vision sites
-    # (A-11 / codec_vision territory, not A-12).
+    # the ONLY inline /chat/completions occurrences left are the vision sites
+    # (A-11 / codec_vision territory, not A-12). D5 / SR-46 moved the webcam
+    # vision POST to routes/media.py — pinning the count at 4 here + the
+    # webcam site is covered by the A-12 allowlist in test_a12_invariant.
     src = (REPO / "codec_dashboard.py").read_text()
-    assert src.count("/chat/completions") == 5
+    assert src.count("/chat/completions") == 4
 
 
 def test_chat_handler_uses_codec_llm_stream():
